Should work as intended now

+++ scripts/deploy.sh
@@ -1,11 +1,17 @@
 #!/bin/bash
 cd /home/ec2-user/app
 
-# Pull the latest code (optional, as CodeDeploy will do this)
-# git pull origin main
-
 # Build and start the containers
-docker-compose up --build -d
+DOCKER_COMPOSE_PATH="/usr/local/bin/docker-compose"
+
+# Check if docker-compose exists at the specified path
+if [ ! -x "$DOCKER_COMPOSE_PATH" ]; then
+    echo "Error: docker-compose not found at $DOCKER_COMPOSE_PATH"
+    exit 1
+fi
+
+#take down current version
+$DOCKER_COMPOSE_PATH up --build -d
 
 # Remove unused images and volumes (optional)
 docker image prune -f
